Modularized setupDB.js
